#33 Add rope companion build for `AnyString`
